+++ saleor/graphql/extensions/mutations.py
@@ -2,7 +2,7 @@ import graphene
 
 from ...extensions.manager import get_extensions_manager
 from ..core.mutations import BaseMutation
-from .types import PluginConfiguration
+from .types import Plugin
 
 
 class ConfigurationItemInput(graphene.InputObjectType):
@@ -12,7 +12,7 @@ class ConfigurationItemInput(graphene.InputObjectType):
     )
 
 
-class PluginConfigurationUpdateInput(graphene.InputObjectType):
+class PluginUpdateInput(graphene.InputObjectType):
     active = graphene.Boolean(
         required=False, description="Indicates whether the plugin should be enabled"
     )
@@ -23,12 +23,12 @@ class PluginConfigurationUpdateInput(graphene.InputObjectType):
     )
 
 
-class PluginConfigurationUpdate(BaseMutation):
-    plugin_configuration = graphene.Field(PluginConfiguration)
+class PluginUpdate(BaseMutation):
+    plugin = graphene.Field(Plugin)
 
     class Arguments:
         id = graphene.ID(required=True, description="ID of plugin to update")
-        input = PluginConfigurationUpdateInput(
+        input = PluginUpdateInput(
             description="Fields required to update a plugin configuration.",
             required=True,
         )
@@ -41,7 +41,7 @@ class PluginConfigurationUpdate(BaseMutation):
     def perform_mutation(cls, root, info, **data):
         plugin_id = data.get("id")
         input = data.get("input")
-        instance = cls.get_node_or_error(info, plugin_id, only_type=PluginConfiguration)
+        instance = cls.get_node_or_error(info, plugin_id, only_type=Plugin)
         manager = get_extensions_manager()
         instance = manager.save_plugin_configuration(instance.name, input)
-        return PluginConfigurationUpdate(plugin_configuration=instance)
+        return PluginUpdate(plugin=instance)