+++ tests/test_product.py
@@ -1,15 +1,11 @@
 import datetime
-import io
-import json
 import os
 from decimal import Decimal
 from unittest.mock import patch
 
 import pytest
 from bs4 import BeautifulSoup, Tag
-from django.core import serializers
 from django.core.exceptions import ValidationError
-from django.core.serializers.base import DeserializationError
 from django.http import JsonResponse
 from django.template.response import TemplateResponse
 from django.urls import reverse
@@ -398,7 +394,9 @@ def test_adding_to_checkout_with_closed_checkout_token(
 
 def test_product_filter_before_filtering(authorized_client, product, category):
     products = (
-        models.Product.objects.all().filter(category__name=category).order_by("-price")
+        models.Product.objects.all()
+        .filter(category__name=category)
+        .order_by("-price_amount")
     )
     url = reverse(
         "product:category", kwargs={"slug": category.slug, "category_id": category.pk}
@@ -411,12 +409,17 @@ def test_product_filter_before_filtering(authorized_client, product, category):
 
 def test_product_filter_product_exists(authorized_client, product, category):
     products = (
-        models.Product.objects.all().filter(category__name=category).order_by("-price")
+        models.Product.objects.all()
+        .filter(category__name=category)
+        .order_by("-price_amount")
     )
     url = reverse(
         "product:category", kwargs={"slug": category.slug, "category_id": category.pk}
     )
-    data = {"minimal_variant_price_min": [""], "minimal_variant_price_max": ["20"]}
+    data = {
+        "minimal_variant_price_amount_min": [""],
+        "minimal_variant_price_amount_max": ["20"],
+    }
 
     response = authorized_client.get(url, data)
 
@@ -502,7 +505,7 @@ def test_product_filter_form(authorized_client, product, category):
     products = (
         models.Product.objects.all()
         .filter(category__name=category)
-        .order_by("-minimal_variant_price")
+        .order_by("-minimal_variant_price_amount")
     )
     url = reverse(
         "product:category", kwargs={"slug": category.slug, "category_id": category.pk}
@@ -521,12 +524,12 @@ def test_product_filter_sorted_by_price_descending(
     products = (
         models.Product.objects.all()
         .filter(category__name=category, is_published=True)
-        .order_by("-minimal_variant_price")
+        .order_by("-minimal_variant_price_amount")
     )
     url = reverse(
         "product:category", kwargs={"slug": category.slug, "category_id": category.pk}
     )
-    data = {"sort_by": "-minimal_variant_price"}
+    data = {"sort_by": "-minimal_variant_price_amount"}
 
     response = authorized_client.get(url, data)
 
@@ -829,99 +832,6 @@ def test_variant_base_price(product, price_override):
     variant.save()
 
     assert variant.base_price == variant.price_override
-
-
-def test_product_json_serialization(product):
-    product.price = Money("10.00", "USD")
-    product.save()
-    data = json.loads(serializers.serialize("json", models.Product.objects.all()))
-    assert data[0]["fields"]["price"] == {
-        "_type": "Money",
-        "amount": "10.00",
-        "currency": "USD",
-    }
-
-
-def test_product_json_deserialization(category, product_type):
-    product_json = """
-    [{{
-        "model": "product.product",
-        "pk": 60,
-        "fields": {{
-            "seo_title": null,
-            "seo_description": "Future almost cup national.",
-            "product_type": {product_type_pk},
-            "name": "Kelly-Clark",
-            "description": "Future almost cup national",
-            "category": {category_pk},
-            "price": {{"_type": "Money", "amount": "35.98", "currency": "USD"}},
-            "minimal_variant_price":
-                {{"_type": "Money", "amount": "35.98", "currency": "USD"}},
-            "publication_date": null,
-            "is_published": true,
-            "attributes": "{{\\"9\\": \\"24\\", \\"10\\": \\"26\\"}}",
-            "updated_at": "2018-07-19T13:30:24.195Z",
-            "is_featured": false,
-            "charge_taxes": true,
-            "tax_rate": "standard"
-        }}
-    }}]
-    """.format(
-        category_pk=category.pk, product_type_pk=product_type.pk
-    )
-    product_deserialized = list(
-        serializers.deserialize("json", product_json, ignorenonexistent=True)
-    )[0]
-    product_deserialized.save()
-    product = models.Product.objects.first()
-    assert product.price == Money(Decimal("35.98"), "USD")
-
-    # same test for bytes
-    product_json_bytes = bytes(product_json, "utf-8")
-    product_deserialized = list(
-        serializers.deserialize("json", product_json_bytes, ignorenonexistent=True)
-    )[0]
-    product_deserialized.save()
-    product = models.Product.objects.first()
-    assert product.price == Money(Decimal("35.98"), "USD")
-
-    # same test for stream
-    product_json_stream = io.StringIO(product_json)
-    product_deserialized = list(
-        serializers.deserialize("json", product_json_stream, ignorenonexistent=True)
-    )[0]
-    product_deserialized.save()
-    product = models.Product.objects.first()
-    assert product.price == Money(Decimal("35.98"), "USD")
-
-
-def test_json_no_currency_deserialization(category, product_type):
-    product_json = """
-    [{{
-        "model": "product.product",
-        "pk": 60,
-        "fields": {{
-            "seo_title": null,
-            "seo_description": "Future almost cup national.",
-            "product_type": {product_type_pk},
-            "name": "Kelly-Clark",
-            "description": "Future almost cup national",
-            "category": {category_pk},
-            "price": {{"_type": "Money", "amount": "35.98"}},
-            "publication_date": null,
-            "is_published": true,
-            "attributes": "{{\\"9\\": \\"24\\", \\"10\\": \\"26\\"}}",
-            "updated_at": "2018-07-19T13:30:24.195Z",
-            "is_featured": false,
-            "charge_taxes": true,
-            "tax_rate": "standard"
-        }}
-    }}]
-    """.format(
-        category_pk=category.pk, product_type_pk=product_type.pk
-    )
-    with pytest.raises(DeserializationError):
-        list(serializers.deserialize("json", product_json, ignorenonexistent=True))
 
 
 def test_variant_picker_data_with_translations(