+++ CHANGELOG.md
@@ -57,6 +57,10 @@ All notable, unreleased changes to this project will be documented in this file.
 - Add an functionality to sort products by their "minimal variant price" - #4416 by @derenio
 - New stripe gateway implementation based on Stripe PaymentIntents API - #4606 by @salwator
 - Service (bot) accounts - backend support - #4689 by @korycins
+- Change AddressValidationRules API - #4655 by @Kwaidan00
+- Refactor account deletion mutations - #4668 by @fowczarek
+- Upgraded django-prices from v1 to v2.1. Currency codes are now locked at 3 characters max by default for consistency. - #4639 by @NyanKiyoshi
+- Drop deprecated fields from api - #4684 by@fowczarek
 
 ## 2.8.0